move enum class for string distance to jaccard distance init
