+++ valentine/algorithms/jaccard_distance/__init__.py
@@ -1,0 +1,9 @@
+from enum import Enum, auto
+
+class StringDistanceFunction(Enum):
+    Levenshtein = auto()
+    DamerauLevenshtein = auto()
+    Jaro = auto()
+    JaroWinkler = auto()
+    Hamming = auto()
+    Exact = auto()
